Resolved issue with distribution-build, and cleaned up a few bits.
Turns out that filename-resolution works completely differently, if the file you're resolving is in a Windows filesystem or inside a JAR's filesystem. I had an issue when executing the distribution .bat, whereby the map JPG couldn't be found inside the JAR. Turns out I had the extension case wrong: the file was named ".jpg", but I was looking for ".JPG". That can be resolved perfectly well on a native Windows filesystem, but *not* when operating in a JAR's "filesystem".